# perf(*): faster string splitting and an iterator