--- a/spec/04-perf/01-rps/01-simple_spec.lua
+++ b/spec/04-perf/01-rps/01-simple_spec.lua
@@ -1,5 +1,5 @@
 local perf = require("spec.helpers.perf")
-local split = require("pl.stringx").split
+local splitn = require("kong.tools.string").splitn
 local utils = require("spec.helpers.perf.utils")
 
 perf.use_defaults()
@@ -8,7 +8,7 @@ local versions = {}
 
 local env_versions = os.getenv("PERF_TEST_VERSIONS")
 if env_versions then
-  versions = split(env_versions, ",")
+  versions = splitn(env_versions, ",")
 end
 
 local LOAD_DURATION = os.getenv("PERF_TEST_LOAD_DURATION") or 30