### Summary

On my testing the `splitn` implemented here is about 10x faster than the Penlight's `stringx.split`.

```
  splitn:  193 ms (new function)
 isplitn:  379 ms (iterator over splitn)
   split: 1638 ms (penlight and kong.tools.string.split as it is alias to penlight)
re.split: 1072 ms (ngx.re)
```

I also tested implementing `isplit` iterator without `splitn` and table allocation, but in my testing that didn't yield better performance.

Microbenchmark (yes, all of them are fast and they may do slightly different splits, but one is clearly the fastest):
```lua
local monotonic_msec = require("resty.core.time").monotonic_msec
local collectgarbage = collectgarbage
local update_time = ngx.update_time
local split1 = require("kong.tools.string").splitn
local split2 = require("kong.tools.string").split
local split3 = require("ngx.re").split
local sleep = ngx.sleep
local print = print

local TEST_SET = {
  "",
  ",",
  ",,",
  "a,,b",
  "a,b,c",
  "cat dog,fish bowl,qwerty,jack,,dog,mermaid,placeholderplaceholderplaceholderplaceholderplaceholderplaceholderplaceholder,wc",
  "cat dog,fish bowl,qwerty,jack,,dog,mermaid,placeholderplaceholderplaceholderplaceholderplaceholderplaceholderplaceholder,wc, 234",
}

local TEST_COUNT = #TEST_SET
local ITERATIONS = 1000000

local s, e
for _ = 1, 2 do
  collectgarbage()
  collectgarbage()
  collectgarbage()

  sleep(1)

  update_time()
  s = monotonic_msec()
  for _ = 1, ITERATIONS do
    for j = 1, TEST_COUNT do
      local _ = split1(TEST_SET[j], ",")
    end
  end
  update_time()
  e = monotonic_msec()
  print("  splitn: ", e - s, " ms")

  collectgarbage()
  collectgarbage()
  collectgarbage()

  sleep(1)

  update_time()
  s = monotonic_msec()
  for _ = 1, ITERATIONS do
    for j = 1, TEST_COUNT do
      local _ = split2(TEST_SET[j], ",")
    end
  end
  update_time()
  e = monotonic_msec()
  print("   split: ", e - s, " ms")

  collectgarbage()
  collectgarbage()
  collectgarbage()

  sleep(1)

  update_time()
  s = monotonic_msec()
  for _ = 1, ITERATIONS do
    for j = 1, TEST_COUNT do
      local _ = split3(TEST_SET[j], ",", "jo")
    end
  end
  update_time()
  e = monotonic_msec()
  print("re.split: ", e - s, " ms")
end
```

### Checklist

- [x] The Pull Request has tests
- [x] A changelog file has been created under `changelog/unreleased/kong` or `skip-changelog` label added on PR if changelog is unnecessary. [README.md](https://github.com/Kong/gateway-changelog/blob/main/README.md)
- [ ] There is a user-facing docs PR against https://github.com/Kong/docs.konghq.com - PUT DOCS PR HERE

### Issue reference

KAG-6825