@bungle 2025-04-16
> do you think it would be possible to add some sort of performance test?

I added microbenchmark at the top. But I don't think that is very good. But it shows some results. Basically it just shows that my implementation probably does less frequent allocations, and most of the speed is from that. Also creates perhaps tighter code that LuaJIT is able to JIT compile better.